+++ utils/tests/utils.go
@@ -114,8 +114,8 @@ func init() {
 	TestPip = flag.Bool("test.pip", false, "Test Pip")
 	TestPipenv = flag.Bool("test.pipenv", false, "Test Pipenv")
 	TestPoetry = flag.Bool("test.poetry", false, "Test Poetry")
-	TestUv     = flag.Bool("test.uv", false, "Test UV")
-	TestNix   = flag.Bool("test.nix", false, "Test Nix")
+	TestUv = flag.Bool("test.uv", false, "Test UV")
+	TestNix = flag.Bool("test.nix", false, "Test Nix")
 	TestConan = flag.Bool("test.conan", false, "Test Conan")
 	TestHelm = flag.Bool("test.helm", false, "Test Helm")
 	TestHuggingFace = flag.Bool("test.huggingface", false, "Test HuggingFace")
@@ -862,6 +862,31 @@ func SetupGitHubActionsEnv(t *testing.T) (cleanup func(), actualOrg, actualRepo 
 	return cleanup, actualOrg, actualRepo
 }
 
+// SetupGitHubActionsEnvForLocalGitMerge enables CI VCS collection with provider/org/repo
+// but clears url/revision/branch CI env vars so local git fallback is exercised.
+func SetupGitHubActionsEnvForLocalGitMerge(t *testing.T) (cleanup func(), actualOrg, actualRepo string) {
+	t.Helper()
+	cleanupBase, actualOrg, actualRepo := SetupGitHubActionsEnv(t)
+
+	var callbacks []func()
+	for _, key := range []string{
+		"GITHUB_SERVER_URL",
+		"GITHUB_SHA",
+		"GITHUB_REF",
+		"GITHUB_REF_NAME",
+		"GITHUB_HEAD_REF",
+	} {
+		callbacks = append(callbacks, tests.SetEnvWithCallbackAndAssert(t, key, ""))
+	}
+
+	return func() {
+		for _, cb := range callbacks {
+			cb()
+		}
+		cleanupBase()
+	}, actualOrg, actualRepo
+}
+
 // ValidateCIVcsPropsOnArtifacts validates that CI VCS properties are set on artifacts.
 func ValidateCIVcsPropsOnArtifacts(t *testing.T, resultItems []utils.ResultItem, expectedProvider, expectedOrg, expectedRepo string) {
 	for _, item := range resultItems {
@@ -957,6 +982,79 @@ func ValidateCIVcsPropsIfPresent(t *testing.T, resultItems []utils.ResultItem, e
 			vals, ok := propertiesMap["vcs.repo"]
 			assert.True(t, ok, "Missing vcs.repo on %s", item.Name)
 			assert.Contains(t, vals, expectedRepo, "Wrong vcs.repo on %s", item.Name)
+		}
+	}
+}
+
+// SetupLocalGitVcsEnv enables VCS property collection and clears CI detection
+// so only local git fallback is exercised.
+func SetupLocalGitVcsEnv(t *testing.T) (cleanup func()) {
+	t.Helper()
+	var callbacks []func()
+
+	for _, key := range []string{
+		"JFROG_CLI_CI_VCS_PROPS_DISABLED", // set to "" to enable
+		"CI", "GITHUB_ACTIONS", "GITHUB_WORKFLOW", "GITHUB_RUN_ID",
+		"GITHUB_REPOSITORY", "GITHUB_REPOSITORY_OWNER",
+		"GITHUB_SERVER_URL", "GITHUB_SHA", "GITHUB_REF", "GITHUB_REF_NAME", "GITHUB_HEAD_REF",
+	} {
+		callbacks = append(callbacks, tests.SetEnvWithCallbackAndAssert(t, key, ""))
+	}
+
+	return func() {
+		for _, cb := range callbacks {
+			cb()
+		}
+	}
+}
+
+// ValidateLocalGitVcsPropsOnArtifacts asserts vcs.url, vcs.revision, vcs.branch on every item.
+func ValidateLocalGitVcsPropsOnArtifacts(t *testing.T, resultItems []utils.ResultItem, expectedURL, expectedRevision, expectedBranch string) {
+	t.Helper()
+	for _, item := range resultItems {
+		propertiesMap := ConvertPropertiesToMap(item.Properties)
+		assertLocalGitProp(t, item.Name, propertiesMap, "vcs.url", expectedURL)
+		assertLocalGitProp(t, item.Name, propertiesMap, "vcs.revision", expectedRevision)
+		if expectedBranch != "" {
+			assertLocalGitProp(t, item.Name, propertiesMap, "vcs.branch", expectedBranch)
+		}
+	}
+}
+
+func assertLocalGitProp(t *testing.T, itemName string, props map[string][]string, key, expected string) {
+	t.Helper()
+	vals, ok := props[key]
+	assert.True(t, ok, "Missing %s on %s", key, itemName)
+	assert.Contains(t, vals, expected, "Wrong %s on %s", key, itemName)
+}
+
+// ValidateNoLocalGitVcsPropsOnArtifacts asserts url/revision/branch are absent.
+func ValidateNoLocalGitVcsPropsOnArtifacts(t *testing.T, resultItems []utils.ResultItem) {
+	t.Helper()
+	for _, item := range resultItems {
+		propertiesMap := ConvertPropertiesToMap(item.Properties)
+		_, hasURL := propertiesMap["vcs.url"]
+		_, hasRev := propertiesMap["vcs.revision"]
+		_, hasBranch := propertiesMap["vcs.branch"]
+		assert.False(t, hasURL, "vcs.url should not be set on %s", item.Name)
+		assert.False(t, hasRev, "vcs.revision should not be set on %s", item.Name)
+		assert.False(t, hasBranch, "vcs.branch should not be set on %s", item.Name)
+	}
+}
+
+// ValidateCIAndLocalGitVcsPropsOnArtifacts asserts CI props plus local git props coexist.
+func ValidateCIAndLocalGitVcsPropsOnArtifacts(t *testing.T, resultItems []utils.ResultItem,
+	expectedProvider, expectedOrg, expectedRepo, expectedURL, expectedRevision, expectedBranch string) {
+	t.Helper()
+	for _, item := range resultItems {
+		propertiesMap := ConvertPropertiesToMap(item.Properties)
+		assertLocalGitProp(t, item.Name, propertiesMap, "vcs.provider", expectedProvider)
+		assertLocalGitProp(t, item.Name, propertiesMap, "vcs.org", expectedOrg)
+		assertLocalGitProp(t, item.Name, propertiesMap, "vcs.repo", expectedRepo)
+		assertLocalGitProp(t, item.Name, propertiesMap, "vcs.url", expectedURL)
+		assertLocalGitProp(t, item.Name, propertiesMap, "vcs.revision", expectedRevision)
+		if expectedBranch != "" {
+			assertLocalGitProp(t, item.Name, propertiesMap, "vcs.branch", expectedBranch)
 		}
 	}
 }
